requirements.txt
aio_pika==7.2.0
Faker==18.9.0
pymongo[srv]==3.12.1
python-dotenv==0.19.2
python-dotenv==1.0.0